requirements.txt
# Web frameworks
aiohttp==3.8.5
httpx==0.25.0
fastapi==0.115.6
uvicorn==0.34.0
backoff==2.2.1

# Task managment
arq==0.25.0
redis==5.2.1

# DI
dependency-injector==4.44.0

# configuration
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Serializers
orjson==3.10.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0

# CLI
typer==0.15.1

# Web utils
furl==2.1.3

#tests
pytest==8.3.4
mimesis==18.0.0
pytest-mock==3.14.0